% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_helper_functions.R
\name{bom_date_to_ymd}
\alias{bom_date_to_ymd}
\title{Convert Date objects to a data frame of integer year, month, day values}
\usage{
bom_date_to_ymd(dates)
}
\arguments{
\item{dates}{A vector of \code{Date} objects}
}
\value{
A data frame with columns year, month, day
}
\description{
This is a helper for other package functions.
}
