% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper_functions.R
\name{bom_db_get_file_name}
\alias{bom_db_get_file_name}
\title{Extract file names from full file paths}
\usage{
bom_db_get_file_name(paths)
}
\arguments{
\item{paths}{A character vector of one or more file paths.}
}
\value{
A vector of file names
}
\description{
This is a helper used by other package functions. It extracts the
file name part of each full path in the \code{paths} argument.
}
