% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{bom_db_is_connected}
\alias{bom_db_is_connected}
\title{Check if a database connection is valid and open}
\usage{
bom_db_is_connected(con)
}
\arguments{
\item{con}{A database connection object, e.g. created using
\code{DBI::dbConnect()} or \code{pool::dbPool()}.}
}
\value{
\code{TRUE} if the connection is valid and currently open;
  \code{FALSE} otherwise.
}
\description{
This function tests that a connection object is both valid and pointing to an
open connection. R packages for working with particular databases such as
\code{RPostgres} and \code{RSQLite}, as well as related higher level packages
such as \code{pool} and \code{DBI}, provide a function \code{dbIsValid} to
test if an object is a valid database connection. However, this function only
checks that an object is of the right type (i.e. it \emph{is} a database
connection object) and not that the connection is actually open, so
\code{dbIsValid(mydb)} can return \code{TRUE} even if the connection has been
closed or the object \code{mydb} is pointing to an incorrect database path.
}
