% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper_functions.R
\name{bom_db_station_id}
\alias{bom_db_station_id}
\title{Format BOM weather station numbers as six-character strings}
\usage{
bom_db_station_id(ids)
}
\arguments{
\item{ids}{Vector (integer or character) of one or more station
identifying numbers.}
}
\value{
A character vector of weather station numbers in standard format.
}
\description{
Given a vector of weather station numbers as either integers or text strings,
this function formats each number as a six-character string with leading
zeroes (if required). For example, the integer station number \code{67105}
will be converted to the string \code{'067105'}. This is useful when matching
weather station numbers to BOM data file names because it avoids ambiguity,
e.g. incorrectly matching an integer ID 1005 to a filename that includes
"041005". If the input is a character vector, the function checks that no
elements have more than six characters and then returns a vector with each
element left-padded with \code{'0'}.
}
\examples{
# This will return the vector c("067105", "067108")
bom_db_station_id(c(67105, 67108))

}
