% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper_functions.R
\name{bom_db_tidy_data}
\alias{bom_db_tidy_data}
\title{Convert raw BOM weather station data into a standard format}
\usage{
bom_db_tidy_data(dat.raw)
}
\arguments{
\item{dat.raw}{A data frame of records in raw BOM weather station data format.}
}
\value{
A data frame in the standard format used with the CERMB PostgreSQL
  weather database, with the attribute \code{'datatype'} set to either 'aws'
  or 'synoptic'.
}
\description{
This function takes a data frame of records read from a BOM weather station
data file (either AWS or synoptic) and reformats it to the column names and
data types used in the PostgreSQL weather database. BOM data files use
extremely long column names with slight differences between AWS and synoptic
data. The input year, month, and day columns for local and standard time are
combined into a date_local and date_standard columns. The input hour and
minute columns for local and standard time are kept separate as: hour_local,
hour_standard, min_local and min_standard. The mapping of input column names
to standard names is defined in the lookup table \code{COLUMN_LOOKUP} that is
part of this package.
}
