% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper_functions.R
\name{bom_extract_station_numbers}
\alias{bom_extract_station_numbers}
\title{Extract integer weather station identifiers from file names}
\usage{
bom_extract_station_numbers(filenames)
}
\arguments{
\item{filenames}{A character vector of paths and/or file names of BOM weather
station data files.}
}
\value{
A character vector of weather station numbers. Each element will include
  any leading zeroes (e.g. "067105").
}
\description{
Given a vector of one or more paths and file names for BOM weather station
dat files, this function extracts the integer station identifiers, including
any leading zeroes, as a vector of character strings (e.g. "067105"). The
function relies on BOM conventions for data file names. The returned vector
can be converted to a vector of integer IDs using the base R
\code{as.integer} function.
}
