% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_helper_functions.R
\name{bom_is_zip_file}
\alias{bom_is_zip_file}
\title{Check if a file name refers to a zip file}
\usage{
bom_is_zip_file(filenames)
}
\arguments{
\item{filenames}{A character vector of paths and/or file names.}
}
\value{
A logical vector in which an element will be \code{TRUE} if the
  corresponding input file name referred to a zip file.
}
\description{
This is a helper used by other package functions. It simply checks whether
the file name has a '.zip' extension (case-insensitive).
}
