% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_helper_functions.R
\name{bom_ymd_to_date}
\alias{bom_ymd_to_date}
\title{Convert integer year, month, day values to Date objects.}
\usage{
bom_ymd_to_date(year, month, day)
}
\arguments{
\item{year}{Either a vector of four digit year numbers, or a data frame or matrix with
column names year, month, day (all lower case).}

\item{month}{A vector of month numbers with the same length as \code{year}. Ignored if
\code{year} is a matrix or data frame.}

\item{day}{A vector of day numbers with the same length as \code{year}. Ignored if
\code{year} is a matrix or data frame.}
}
\value{
A vector of \code{Date} objects.
}
\description{
This is a helper for other package functions. If year is a data frame or
matrix it is assumed to have columns year, month and day. Otherwise three
equal-length vectors are expected.
}
